Rename package cook.scheduler.client -> cookclient
